+++ libbeat/docs/processors-list.asciidoc
@@ -208,7 +208,7 @@ ifndef::no_translate_sid_processor[]
 include::{libbeat-processors-dir}/translate_sid/docs/translate_sid.asciidoc[]
 endif::[]
 ifndef::no_urldecode_processor[]
-include::{libbeat-processors-dir}/actions/docs/urldecode.asciidoc[]
+include::{libbeat-processors-dir}/urldecode/docs/urldecode.asciidoc[]
 endif::[]
 
 //# end::processors-include[]
