+++ libbeat/processors/urldecode/urldecode_test.go
@@ -15,7 +15,7 @@
 // specific language governing permissions and limitations
 // under the License.
 
-package actions
+package urldecode
 
 import (
 	"testing"
